refactor: else if 제거, 패키지 구조 변경, getter 네이밍 변경
