+++ src/main/java/racingcar/domain/Number.java
@@ -1,4 +1,4 @@
-package racingcar.infra;
+package racingcar.domain;
 
 public interface Number {
     int generate();
