+++ src/test/java/racingcar/MoveCommandTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.MoveCommand;
 import racingcar.domain.MoveType;
-import racingcar.infra.MoveCommandImpl;
 import racingcar.utils.TestNumber;
 
 import java.util.stream.Stream;
@@ -20,7 +19,7 @@ public class MoveCommandTest {
     @MethodSource("provideValidMoveNumber")
     @DisplayName("0 ~ 3 일 경우 정지하고, 4 ~ 9 일 경우 앞으로 가라고 명령하는 지 테스트")
     public void testMoveCommandUsingValidNumber(int number, MoveType expected) {
-        MoveCommand moveCommand = new MoveCommandImpl(new TestNumber(number));
+        MoveCommand moveCommand = new MoveCommand(new TestNumber(number));
 
         assertThat(moveCommand.command()).isEqualTo(expected);
     }
@@ -29,7 +28,7 @@ public class MoveCommandTest {
     @ValueSource(ints = {-1, 10})
     @DisplayName("0~9를 벗어나는 숫자일 경우 예외 발생")
     public void testMoveCommandUsingInvalidNumber(int number) {
-        MoveCommand moveCommand = new MoveCommandImpl(new TestNumber(number));
+        MoveCommand moveCommand = new MoveCommand(new TestNumber(number));
 
         assertThatThrownBy(moveCommand::command).isInstanceOf(RuntimeException.class);
     }
